+++ src/shared/ui/button/Button.stories.tsx
@@ -1,15 +1,15 @@
 import { Meta, StoryObj } from '@storybook/react';
 import React from 'react';
-import Button from './Button';
 import IcBookmark from '@/shared/assets/svg/IcBookmark.svg?react';
 import IcChat from '@/shared/assets/svg/IcChat.svg?react';
 import IcChatGithub from '@/shared/assets/svg/IcChatGithub.svg?react';
 import IcChatStart from '@/shared/assets/svg/IcChatStart.svg?react';
-import IcClose from '@/shared/assets/svg/Icclose.svg?react';
+import IcClose from '@/shared/assets/svg/IcClose.svg?react';
 import IcList from '@/shared/assets/svg/IcList.svg?react';
 import IcLoginGithub from '@/shared/assets/svg/IcLoginGithub.svg?react';
 import IcPhoto from '@/shared/assets/svg/IcPhoto.svg?react';
 import IcSend from '@/shared/assets/svg/IcSend.svg?react';
+import { Button } from '@/shared/ui';
 import { StoryContainer, SectionContainer, RowContainer } from '@/shared/ui/storybook/story.style';
 
 const meta: Meta<typeof Button> = {
